backend/pyproject.toml
[tool.poetry]
name = "artist_registry"
version = "0.1.0"
description = ""
authors = ["Jan Miksik"]

[tool.poetry.dependencies]
python = "^3.11"
asgiref = "^3.8.1"
django = "^5.2.0"
djangorestframework = "^3.16.1"
gunicorn = "20.1.0"
psycopg2-binary = "2.9.9"
pytz = "2023.3.post1"
sqlparse = "0.4.4"
django-cors-headers = "^4.3.1"
django-storages = "^1.14.2"
boto3 = "^1.34.29"
python-dotenv = "^1.0.1"
requests = "^2.31.0"
arweave-python-client = "^1.0.19"
node-py = "^0.0.25"
pillow = "^10.2.0"
weaviate-client = "4.9.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"